Use api endpoints for market guides
